Moved all classes to different packages depending on their purpose
