+++ app/src/main/java/com/example/moviereviewer/DataClasses/Movie.java
@@ -1,4 +1,4 @@
-package com.example.moviereviewer;
+package com.example.moviereviewer.DataClasses;
 
 public class Movie {
     private String title;
