+++ app/src/main/java/com/example/moviereviewer/DataClasses/Rating.java
@@ -1,4 +1,4 @@
-package com.example.moviereviewer;
+package com.example.moviereviewer.DataClasses;
 
 public class Rating {
     private int plotRating;
