+++ app/src/main/java/com/example/moviereviewer/Dialogs/CustomAlertDialog.java
@@ -1,15 +1,14 @@
-package com.example.moviereviewer;
+package com.example.moviereviewer.Dialogs;
 
 import android.app.Activity;
 import android.app.Dialog;
 import android.content.Context;
-import android.graphics.Color;
-import android.graphics.drawable.ColorDrawable;
 import android.view.View;
-import android.view.Window;
 import android.widget.TextView;
 
 import com.bumptech.glide.Glide;
+import com.example.moviereviewer.DataClasses.Movie;
+import com.example.moviereviewer.R;
 import com.google.android.material.imageview.ShapeableImageView;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
