+++ app/src/main/java/com/example/moviereviewer/Dialogs/VerificationDialog.java
@@ -1,4 +1,4 @@
-package com.example.moviereviewer;
+package com.example.moviereviewer.Dialogs;
 
 import android.app.Activity;
 import android.app.Dialog;
@@ -9,6 +9,8 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 
+import com.example.moviereviewer.Dialogs.CustomDialog;
+import com.example.moviereviewer.R;
 import com.google.android.material.progressindicator.LinearProgressIndicator;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
