+++ app/src/main/java/com/example/moviereviewer/Fragments/MoviesFragment.java
@@ -1,21 +1,21 @@
-package com.example.moviereviewer;
+package com.example.moviereviewer.Fragments;
 
-import android.annotation.SuppressLint;
 import android.os.Bundle;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import android.view.LayoutInflater;
-import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.airbnb.lottie.LottieAnimationView;
+import com.example.moviereviewer.Dialogs.CustomAlertDialog;
+import com.example.moviereviewer.DataClasses.Movie;
+import com.example.moviereviewer.Adapters.MovieAdapter;
+import com.example.moviereviewer.R;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
