+++ app/src/main/java/com/example/moviereviewer/HelperClasses/StormAppGlideModule.java
@@ -1,4 +1,4 @@
-package com.example.moviereviewer;
+package com.example.moviereviewer.HelperClasses;
 
 import android.content.Context;
 
@@ -8,7 +8,6 @@ import com.bumptech.glide.Glide;
 import com.bumptech.glide.Registry;
 import com.bumptech.glide.annotation.GlideModule;
 import com.bumptech.glide.module.AppGlideModule;
-import com.example.moviereviewer.FirebaseImageLoader;
 import com.google.firebase.storage.StorageReference;
 
 import java.io.InputStream;
