--- app/src/main/java/com/example/moviereviewer/SeriesFragment.java
@@ -1,25 +1,0 @@
-package com.example.moviereviewer;
-
-import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-public class SeriesFragment extends Fragment {
-
-    public SeriesFragment() { }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_series, container, false);
-    }
-}
