+++ app/src/main/res/layout/fragment_add_movie.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddMovieFragment"
+    tools:context=".Fragments.AddMovieFragment"
     android:background="@color/grayPrimary">
 
     <androidx.constraintlayout.widget.ConstraintLayout
