+++ app/src/main/res/layout/fragment_movies.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout2"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MoviesFragment"
+    tools:context=".Fragments.MoviesFragment"
     android:background="@color/grayPrimary">
 
     <androidx.recyclerview.widget.RecyclerView
